Merge the exact upstream 6.2.22 tag with ancestry preserved. Resolve Redis core and test conflicts toward upstream behavior while retaining non-conflicting Windows port adaptations and repository-owned MinGW wrappers.

Keep the customized jemalloc 5.2.1 QFork tree unchanged, omit upstream's separate allocator tree, retain 5.0.14 version metadata during integration, and record the first hiredis-layer red build checkpoint in PORTING_PROGRESS.md.

+++ deps/lua/src/lua_struct.c
@@ -313,7 +313,7 @@ static int b_unpack (lua_State *L) {
       case 'b': case 'B': case 'h': case 'H':
       case 'l': case 'L': case 'T': case 'i':  case 'I': {  /* integer types */
         int issigned = islower(opt);
-        lua_Number res = getinteger(data+pos, h.endian, issigned, (int)size);
+        lua_Number res = getinteger(data+pos, h.endian, issigned, size);
         lua_pushnumber(L, res); n++;
         break;
       }
